[SPIFFY-201] Fake long polling for now

+++ src/main/java/io/spiffy/website/cache/Poll.java
@@ -8,5 +8,5 @@ import java.io.Serializable;
 public class Poll implements Serializable {
     private static final long serialVersionUID = 7397770425072264118L;
 
-    private int notifications;
+    private long notifications;
 }
